requirements.txt
langchain
langchain-openai
langchain-chroma
langchain-huggingface
langchain-text-splitters
chromadb
sentence-transformers
pypdf
python-dotenv
streamlit